Moved rpm.py and added source URL.
