+++ backend/unittests/import_observations/test_types.py
@@ -1,6 +1,5 @@
 from unittest import TestCase
 
-from application.import_observations.rpm import RpmVersion
 from application.import_observations.types import ExtendedSemVer
 
 
@@ -38,11 +37,6 @@ class TestExtendedSemVer(TestCase):
 
         ext_semver = ExtendedSemVer.parse("a:1.2")
         self.assertEqual(ext_semver, None)
-
-        rpm_ver = RpmVersion.from_string("1:21.0.6.0.7-1.el9")
-        self.assertEqual(rpm_ver.epoch, 1)
-        self.assertEqual(rpm_ver.version, "21.0.6.0.7")
-        self.assertEqual(rpm_ver.release, "1.el9")
 
     def test_eq(self):
         ext_semver1 = ExtendedSemVer.parse("1.2.3-alpha")
@@ -109,10 +103,6 @@ class TestExtendedSemVer(TestCase):
         ext_semver1 = ExtendedSemVer.parse("1:1.2.3-beta")
         ext_semver2 = ExtendedSemVer.parse("1:1.2.3-alpha")
         self.assertTrue(ext_semver1 > ext_semver2)
-
-        rpm_ver1 = RpmVersion.from_string("1:21.0.6.0.7-1.el9")
-        rpm_ver2 = RpmVersion.from_string("1:21.0.1.0.12-2.el9")
-        self.assertTrue(rpm_ver1 > rpm_ver2)
 
     def test_ge(self):
         ext_semver1 = ExtendedSemVer.parse("1.2.3-alpha")
